chore: move generated graphql files
