+++ .gitattributes
@@ -1,1 +1,1 @@
-src/generated/* linguist-vendored
+generated/* linguist-vendored
